chore: breaks down test infra to several files
